// Frontend/src/app/core/services/aminety-service/models/AmenityListResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AmenityResponseDto } from './AmenityResponseDto';
import type { PaginationDto } from './PaginationDto';
export type AmenityListResponseDto = {
    data?: Array<AmenityResponseDto> | null;
    pagination?: PaginationDto;
};