// Frontend/src/app/core/services/aminety-service/models/UpdateAmenityRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateAmenityRequestDto = {
    name?: string | null;
    slotTypeId?: string | null;
    statusId?: string | null;
    location?: string | null;
    rules?: string | null;
    imageUrl?: string | null;
};